requirements.txt
mutagen
pillow
musicbrainzngs
tifffile
requests
beautifulsoup4
tqdm
concurrent-log-handler
chardet
pyvips-binary